Moves trial pages to lab directory
